% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_nll_gpd}
\alias{compute_nll_gpd}
\title{Computes the negative log likelihood for the GPD}
\usage{
compute_nll_gpd(y, scale, shape)
}
\description{
No threshold parameter
}
\details{
scale has same length as y and shape has length 1
}

