% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_pd1_scale_nll_gpd}
\alias{compute_pd1_scale_nll_gpd}
\title{Computes the first partial derivative (scale) of the negative log likelihood for the GPD}
\usage{
compute_pd1_scale_nll_gpd(y, scale, shape)
}
\arguments{
\item{y}{numeric vector of the data}

\item{scale}{numeric vector of the scape parameters}

\item{shape}{double}
}
\description{
No threshold parameter
}
\details{
scale has same length as y and shape has length 1
}

