% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convexMinorant}
\alias{convexMinorant}
\title{Computes the convex minorant of a polygon.}
\usage{
convexMinorant(x, y)
}
\arguments{
\item{x, y}{the coordinates of the polygon}
}
\value{
vector of the y-coordinates of the convex minorant
}
\description{
Computes the convex minorant of a polygon.
}

