% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedICM.R
\name{FitIsoScaleFixedICM}
\alias{FitIsoScaleFixedICM}
\title{Isotonic estimation (using an adapted version of the ICM algorithm)}
\usage{
FitIsoScaleFixedICM(y, start, shape, max_repetitions = 1e+05)
}
\arguments{
\item{y}{numeric vector of the data}

\item{start}{Numeric vector of the initial scale parameters (will be forced to be admissible)}

\item{shape}{double}

\item{max_repetitions}{Maximal number of repetitions}
}
\value{
isotonic scale parameter estimate and deviance
}
\description{
Isotonic estimation (using an adapted version of the ICM algorithm)
}

