% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FitIsoScaleFixedPG}
\alias{FitIsoScaleFixedPG}
\title{Isotonic estimation (using a projected gradient method)}
\usage{
FitIsoScaleFixedPG(y, scale, shape, max_repetitions = 100000L)
}
\arguments{
\item{y}{numeric vector of the data}

\item{scale}{numeric vector of the scape parameters}

\item{shape}{double}

\item{max_repetitions}{Maximal number of repetitions}
}
\value{
isotonic scale parameter estimate and deviance
}
\description{
Isotonic estimation (using a projected gradient method)
}
\note{
up to now only for positive shape parameters
}

