% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gpd_isotonic_scale_projected_gradient}
\alias{gpd_isotonic_scale_projected_gradient}
\title{Isotonic estimation (using a projected grdient method)}
\usage{
gpd_isotonic_scale_projected_gradient(y, scale, shape, max_iterations)
}
\value{
isotonic scale parameter estimate and deviance
}
\description{
Isotonic estimation (using a projected grdient method)
}
\note{
up to now only for positive shape parameters
}

