% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gpd_scale_isotonic_fit}
\alias{gpd_scale_isotonic_fit}
\title{Isotonic estimation (using an adapted version of the ICM algorithm)}
\usage{
gpd_scale_isotonic_fit(y, start, shape, max_repetitions = 100000L)
}
\arguments{
\item{y}{numeric vector of the data}

\item{start}{Numeric vector of the initial scale parameters (will be forced to be admissible)}

\item{shape}{double}

\item{max_repetitions}{Maximal number of repetitions}
}
\value{
isotonic scale parameter estimate and deviance
}
\description{
Isotonic estimation (using an adapted version of the ICM algorithm)
}

