% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GreatestConvexMinorant}
\alias{GreatestConvexMinorant}
\title{Computes the convex minorant of a vector.}
\usage{
GreatestConvexMinorant(x, y)
}
\arguments{
\item{x, y}{Vector of x and y values}
}
\value{
x.knots, y.knots, y.slopes and the left derivative at all x values
}
\description{
Computes the convex minorant of a vector.
}

