\name{IcmGpd-package}
\alias{IcmGpd-package}
\alias{IcmGpd}
\docType{package}
\title{
Isotonic scale estimation of the GPD
}
\description{
Methods for isotonic estimation based on projected gradient and adapted iterative convex minorant algorithm
}
\details{
\tabular{ll}{
Package: \tab IcmGpd\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-10-26\cr
License: \tab What Licence is it under ?\cr
}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
M. Roth

Maintainer: M. Roth <marsinaus@gmail.com>
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
%\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%% ~~ simple examples of the most important functions ~~
%}
