% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedICM.R
\name{LeftDerivativeGCM}
\alias{LeftDerivativeGCM}
\title{Left GreatestConvexMinorant}
\usage{
LeftDerivativeGCM(x, y)
}
\arguments{
\item{x}{Vector of x values}

\item{y}{Vector of y values}
}
\description{
Left GreatestConvexMinorant
}

