% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MakeScaleAdmissible}
\alias{MakeScaleAdmissible}
\title{Ensure admissibility of the GPD scale parameter
up to now only for positive shape parameters}
\usage{
MakeScaleAdmissible(scale, y, shape)
}
\arguments{
\item{scale}{numeric vector of the scape parameters}

\item{y}{numeric vector of the data}

\item{shape}{double}
}
\value{
isotonic scale parameter that does not violate the GPD constraint
}
\description{
Ensure admissibility of the GPD scale parameter
up to now only for positive shape parameters
}

