% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compute_pd1_scale_nll_gpd}
\alias{compute_pd1_scale_nll_gpd}
\title{Computes the first partial derivative (scale) of the negative log likelihood for the GPD}
\usage{
compute_pd1_scale_nll_gpd(y, scale, shape)
}
\description{
No threshold parameter
}
\details{
scale has same length as y and shape has length 1
}

