% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FitIsoScaleGPD}
\alias{FitIsoScaleGPD}
\title{Estimation of GPD parameters with constant shape parameter and non-decreasing
scale parameter}
\usage{
FitIsoScaleGPD(y, min_shape, max_shape, by = 0.01,
  max_repetitions = 100000L)
}
\arguments{
\item{y}{numeric vector of the data}

\item{min_shape}{double minimum shape value}

\item{max_shape}{double maximum shape value}

\item{by}{double step size for the profile likelihood}

\item{max_repetitions}{Maximal number of repetitions}
}
\value{
isotonic scale parameter estimate and deviance
}
\description{
Estimation of GPD parameters with constant shape parameter and non-decreasing
scale parameter
}

