% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gpd_scale_isotonic_fit}
\alias{gpd_scale_isotonic_fit}
\title{Isotonic estimation (using an adapted version of the ICM algorithm)}
\usage{
gpd_scale_isotonic_fit(y, start, shape)
}
\value{
isotonic scale parameter estimate and deviance
}
\description{
Isotonic estimation (using an adapted version of the ICM algorithm)
}

