% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gpd_scale_isotonic_fit}
\alias{gpd_scale_isotonic_fit}
\title{Isotonic estimation (using an adapted version of the ICM algorithm)}
\usage{
gpd_scale_isotonic_fit(y, start, shape)
}
\value{
isotonic scale parameter estimate and deviance
}
\description{
Isotonic estimation (using an adapted version of the ICM algorithm)
}

