% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_gpd_admissible}
\alias{make_gpd_admissible}
\title{Ensure admissibility of the GPD scale parameter
up to now only for positive shape parameters}
\usage{
make_gpd_admissible(scale, y, shape)
}
\value{
isotonic scale parameter that does not violate the GPD constraint
}
\description{
Ensure admissibility of the GPD scale parameter
up to now only for positive shape parameters
}

